Drop custom tigera operator patches
The Tigera operator expects the ImageSet specs to contain the original
image names, not the custom ones. We'll drop our custom patches and
update the integration tests accordingly.

We need to remove "calico-ctl" from the image spec as the Tigera
operator does not expect it and is going to throw an error.

While at it, we'll avoid using the "match" expression since it's
unsupported on 22.04.